Getting started with field parsing.
The parsing is now working when a .csv file is opened. It currently prints out the parsed fields (This is temporary behavior). I also still need to get the parsing to happen after the user presses OK in the file selection window, not after they choose a .csv in the JFileChooser.