CircleGauge is now usable
The CircleGauge is now usable! Select a numeric data element in the guage creator (You'll know if its numeric if it displays a maximum value, e.x distance) and you will be able to view the fields. I also got the color ranges working. The only things left are the audio alarm and a bug where the gauge doesn't resize with the window, then it will be complete.